+++ packages/zambdas/src/shared/practitioner/helpers.ts
@@ -1,5 +1,6 @@
 import Oystehr from '@oystehr/sdk';
 import { Appointment, Encounter, Location, Patient, Practitioner, PractitionerRole, Resource } from 'fhir/r4b';
+import { PRACTITIONER_CODINGS } from 'utils';
 import { EncounterPackage } from './types';
 
 // called from assign-practitioner and unassign-practitioner zambdas
@@ -86,3 +87,21 @@ export const getVisitResources = async (
     practitionerRole,
   };
 };
+
+export function getAttendingPractitionerId(encounter: Encounter): string {
+  const practitionerId = encounter.participant
+    ?.find(
+      (participant) =>
+        participant.type?.find(
+          (type) =>
+            type.coding?.some(
+              (c) =>
+                c.system === PRACTITIONER_CODINGS.Attender[0].system && c.code === PRACTITIONER_CODINGS.Attender[0].code
+            )
+        )
+    )
+    ?.individual?.reference?.replace('Practitioner/', '');
+
+  if (!practitionerId) throw Error('Attending practitioner not found');
+  return practitionerId;
+}
